LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/container-units-svglength-expected.txt

CONSOLE MESSAGE: Error: Invalid value for <rect> attribute width="calc(10cqmin + 10cqmax)"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute height="calc(10cqw + 3px)"

PASS unitType with container-relative units
FAIL cqw,cqh can be resolved The operation is not supported.
FAIL cqi,cqb can be resolved The operation is not supported.
FAIL cqmin,cqmax can be resolved The operation is not supported.
FAIL calc() with container-relative units can be resolved assert_equals: expected 35 but got 0
FAIL Can modify value with container-relative units The operation is not supported.
FAIL CSS Container Queries Test: container-relative units in SVGLength assert_approx_equals: expected 15 +/- 1 but got 0
